tools/requirements.txt
reportlab>=4.1.0
Pillow==11.1.0  # Match version from main package
